package/gcc: bump 15.x series to 15.2.0
Announcement:
https://gcc.gnu.org/pipermail/gcc-announce/2025/000189.html

GCC 15.2 is a bug-fix release from the GCC 15 branch
containing important fixes for regressions and serious bugs in
GCC 15.1 with more than 123 bugs fixed since the previous release.

For the list of bugs fixed in this version, see:
https://gcc.gnu.org/gcc-15/changes.html#15.2

One patch is renamed to match the filename generated by "git
format-patch" out of the box.

Signed-off-by: Thomas Petazzoni <thomas.petazzoni@bootlin.com>
[Julien: add link to the bug list in commit log]
Signed-off-by: Julien Olivain <ju.o@free.fr>