package/gcc-bare-metal: bump to version 15.2.0 to fix build
Since Buildroot commit [1] "package/gcc: bump 15.x series to 15.2.0",
gcc-bare-metal fails to download its sources. For example using the
commands:

	make zynqmp_kria_kv260_defconfig
	make source

fails with error:

	ERROR: No hash found for gcc-15.1.0.tar.xz

This is because the gcc-bare-metal package reuses the hash file and
the 15.x package patch directory from the gcc package (which was
bumped to 15.2.0), but still defines its _VERSION to 15.1.0.

This commit fixes this issue by bumping gcc-bare-metal to 15.2.0.

Fixes:
https://autobuild.buildroot.org/results/eda879fb8e1ccd696f01927087713b88a4b1a2c5
https://gitlab.com/buildroot.org/buildroot/-/jobs/11026775160

[1] https://gitlab.com/buildroot.org/buildroot/-/commit/b1dd235ad0dd91e1919a167a17cc9cd20a3695b2

Cc: Ibai Erkiaga <ibai.erkiaga-elorza@amd.com>
Cc: Neal Frager <neal.frager@amd.com>
Cc: Thomas Petazzoni <thomas.petazzoni@bootlin.com>
Signed-off-by: Julien Olivain <ju.o@free.fr>
Acked-by: Thomas Petazzoni <thomas.petazzoni@bootlin.com>
Acked-by: Neal Frager <neal.frager@amd.com>
Signed-off-by: Thomas Petazzoni <thomas.petazzoni@bootlin.com>