package/binutils-bare-metal: bump to binutils 2.45
This patch bumps the binutils-bare-metal package to version 2.45. In the
process, it also decouples binutils-bare-metal from the binutils package
since the binutils package does not yet have support for version 2.45.

With the decoupling, the 3 downstream binutils 2.44 patches were evaluated.

Patches not needed for binutils-bare-metal:
0001-sh-conf.patch
0002-poison-system-directories.patch

Patches already included with binutils version 2.45:
0003-objdump-memleak.patch
https://sourceware.org/git/?p=binutils-gdb.git;a=commitdiff;h=ba6ad3a18cb26b79e0e3b84c39f707535bbc344d

Since none of these patches need to be applied to binutils-bare-metal 2.45,
the link to the binutils patch directory is deleted.

Signed-off-by: Neal Frager <neal.frager@amd.com>
Signed-off-by: Julien Olivain <ju.o@free.fr>