configs/galileo: remove defconfig, broken
This defconfig no longer builds:

  https://gitlab.com/buildroot.org/buildroot/-/jobs/11012273310

It started to break not long ago, about 1 month ago, most likely due a
toolchain update:

  https://gitlab.com/buildroot.org/buildroot/-/jobs?name=T%3Dgalileo&kind=BUILD

However, it is build a very old 3.14, which is anyway unmaintained,
and issues with newer toolchains are going to get worse and worse.

Unless a fix is provided, this defconfig is going to be removed.

Cc: Nicholas Sielicki <sielicki@yandex.com>
Signed-off-by: Thomas Petazzoni <thomas.petazzoni@bootlin.com>
Signed-off-by: Arnout Vandecappelle <arnout@rnout.be>
(cherry picked from commit ce76a11630241ed6560b13e5c55686e8441bebd0)
Signed-off-by: Titouan Christophe <titouan.christophe@mind.be>