package/go: security bump to version 1.23.12
go1.23.11 (released 2025-07-08) includes security fixes to the go command,
as well as bug fixes to the compiler, the linker, and the runtime.
See:
- https://go.dev/doc/devel/release#go1.23.11
- https://github.com/golang/go/issues?q=milestone%3AGo1.23.11+label%3ACherryPickApproved
The 1.23.11 version fixes:
- https://www.cve.org/CVERecord?id=CVE-2025-4674

go1.23.12 (released 2025-08-06) includes security fixes to the database/sql
and os/exec packages, as well as bug fixes to the runtime.
See:
- https://go.dev/doc/devel/release#go1.23.12
- https://github.com/golang/go/issues?q=milestone%3AGo1.23.12+label%3ACherryPickApproved
The 1.23.12 version fixes:
- https://www.cve.org/CVERecord?id=CVE-2025-47907

Signed-off-by: Peter Korsgaard <peter@korsgaard.com>
[Julien: add links to release notes, fixed issues and CVEs]
Signed-off-by: Julien Olivain <ju.o@free.fr>